Add listing of primary chunk replica node
This change adds a function to the test script that shows the primary
replica node for a node. The primary replica node is the node that is
used for queries.

A query fails if it involves a chunk that has a primary replica node
that is currently unavailable. This is why a SELECT query in the test
fails the first time when the primary replica node is down. However,
the same query succeeds later because the unavailable node is no
longer the primary for any chunks.